+++ lib/ui/map2/Map2HomeFilters.dart
@@ -8,6 +8,7 @@ import 'package:illinois/ext/Dining.dart';
 import 'package:illinois/ext/Explore.dart';
 import 'package:illinois/ext/LaundryRoom.dart';
 import 'package:illinois/ext/MTD.dart';
+import 'package:illinois/ext/Places.dart';
 import 'package:illinois/model/Building.dart';
 import 'package:illinois/model/Dining.dart';
 import 'package:illinois/model/Explore.dart';
@@ -20,6 +21,7 @@ import 'package:illinois/ui/map2/Map2HomePanel.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:rokwire_plugin/model/event2.dart';
 import 'package:rokwire_plugin/model/explore.dart';
+import 'package:rokwire_plugin/model/places.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
@@ -49,6 +51,7 @@ class Map2Filter {
       case Map2ContentType.LaundryRooms:         return Map2LaundryRoomsFilter.defaultFilter();
       case Map2ContentType.BusStops:             return Map2BusStopsFilter.defaultFilter();
       case Map2ContentType.Therapists:           return null;
+      case Map2ContentType.StoriedSites:         return Map2StoriedSitesFilter.defaultFilter();
       case Map2ContentType.MyLocations:          return Map2MyLocationsFilter.defaultFilter();
       default: return null;
     }
@@ -494,6 +497,90 @@ class Map2BusStopsFilter extends Map2Filter {
     }
     if ((filteredExplores != null) && descriptionMap.isNotEmpty)  {
       String buildingsKey = Localization().getStringEx('panel.map2.filter.bus_stops.text', 'Bus Stops');
+      String buildingsValue = filteredExplores.length.toString();
+      descriptionMap[buildingsKey] = <String>[buildingsValue];
+    }
+    return descriptionMap;
+  }
+}
+
+class Map2StoriedSitesFilter extends Map2Filter {
+  LinkedHashSet<String> tags = LinkedHashSet<String>();
+  bool onlyVisited = false;
+
+  Map2StoriedSitesFilter._({
+    required this.tags,
+    // ignore: unused_element_parameter
+    this.onlyVisited = false,
+
+    String searchText = '',
+    bool starred = false,
+    Map2SortType? sortType,
+    Map2SortOrder? sortOrder,
+  }) : super._(
+    searchText: searchText,
+    starred: starred,
+    sortType: sortType,
+    sortOrder: sortOrder,
+  );
+
+  factory Map2StoriedSitesFilter.defaultFilter() => Map2StoriedSitesFilter._(
+    tags: LinkedHashSet<String>()
+  );
+
+  @override
+  bool get _hasFilter => ((searchText.isNotEmpty == true) || (onlyVisited == true) || (tags.isNotEmpty == true));
+
+  @override
+  List<Explore> _filter(List<Explore> explores) {
+    String? searchLowerCase = searchText.toLowerCase();
+    List<Explore> filtered = <Explore>[];
+    for (Explore explore in explores) {
+      if ((explore is Place) &&
+          ((searchLowerCase.isNotEmpty != true) || (explore.matchSearchTextLowerCase(searchLowerCase))) &&
+          ((onlyVisited != true) || (explore.isVisited == true)) &&
+          ((tags.isNotEmpty != true) || (explore.matchTags(tags)))
+        ) {
+        filtered.add(explore);
+      }
+    }
+    return filtered;
+  }
+
+  @override
+  LinkedHashMap<String, List<String>> description(List<Explore>? filteredExplores, { List<Explore>? explores }) {
+    LinkedHashMap<String, List<String>> descriptionMap = LinkedHashMap<String, List<String>>();
+    if (searchText.isNotEmpty) {
+      String searchKey = Localization().getStringEx('panel.map2.filter.search.text', 'Search');
+      descriptionMap[searchKey] = <String>[searchText];
+    }
+    if (tags.isNotEmpty) {
+      LinkedHashMap<String, LinkedHashSet<String>> displayTags = tags.displayTags;
+      for (String tagCategory in displayTags.keys) {
+        LinkedHashSet<String>? tagCategoryValues = displayTags[tagCategory];
+        if ((tagCategoryValues != null) && tagCategoryValues.isNotEmpty) {
+          String displayCategory = tagCategory.isNotEmpty ? tagCategory : Localization().getStringEx('panel.map2.filter.tags.text', 'Tags');
+          descriptionMap[displayCategory] = tagCategoryValues.toList();
+        }
+      }
+    }
+    if (onlyVisited) {
+      String visitedKey = Localization().getStringEx('panel.map2.filter.visited.text', 'Visited');
+      descriptionMap[visitedKey] = <String>[];
+    }
+    if (sortType != null) {
+      String sortKey = Localization().getStringEx('panel.map2.filter.sort.text', 'Sort');
+      String sortValue = sortType?.displayTitle ?? '';
+      if (sortValue.isNotEmpty && (sortOrder != null)) {
+        String? sortOrderValue = sortOrder?.displayMnemo;
+        if ((sortOrderValue != null) && sortOrderValue.isNotEmpty) {
+          sortValue += " $sortOrderValue";
+        }
+      }
+      descriptionMap[sortKey] = <String>[sortValue];
+    }
+    if ((filteredExplores != null) && descriptionMap.isNotEmpty)  {
+      String buildingsKey = Localization().getStringEx('panel.map2.filter.storied_sites.text', 'Storied Sites');
       String buildingsValue = filteredExplores.length.toString();
       descriptionMap[buildingsKey] = <String>[buildingsValue];
     }